Lots of changes here, mostly made in order to build JNEAT 1)test.sh
test.sh is a very small bash file which removes non utf8 characters from the java source files in jGraph.  This utility should not be needed again, but I am including it in case we have to revert to the base jNEAT

2)build.xml

This is an ant build configuration to build JNEAT.  This script was taken from: http://blogs.lessthandot.com/index.php/desktopdev/suntech/java/integrating-jneat-into-a-custom-simulati/

blame him if things go wrong.  This bypasses the need for the bash scripts which are supposed to compile everything but really dont.

3) jGraph/*.java

These files are the results of running the bash script(test.sh) on jGraph.  Basically, it should be the jGraph files minus the non-utf8 characters.

This is now a compiling build.